Separate A-star algoryth to a helper file inside behaviors
